Read channel and db configuration from a common config file
In the tools world, my.replica.cnf is available in all home folders
which is not the case when setting up the service standalone on labs.
This patch sets up the config so the app can read both channel and
db config from a common file. A default-config file has been added,
and the actual values are stored in staging.cfg and production.cfg
and ignored by git.