Add various optional filters
In the interest of enabling IFTTT users to do a larger array
of cool things, I've interested a few filters

* hrs - allows user to look at a period of time e.g. is it things
that have trended in the last 24 hrs or things that have trended in
the last hour?
* edits - allows users to set the edit threshold. Maybe they want
less noise or more noise
* editors - allows users to work with use case such as "find an editor
who's putting lots of effort into an article on their own". Could be
a good way for editors to connect with one another
* score - this is a bit more nebulous, but through trial and error
allows more control over the results. We'd need to document this
in the UI, and may decide to exclude it in an initial offering but
I think it's good to have
* title_contains - allows editors to set trends on topics which contain
a title for instance maybe they want to monitor articles such
as "2018 Qatar Total Open" or "The X-Files". it's case insensitive and
powerful.